+++ packages/react-native-contentpass-cmp-onetrust/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @contentpass/react-native-contentpass-cmp-onetrust
 
+## 0.4.3
+
+### Patch Changes
+
+- Fetch Android preference-center data before creating the adapter.
+
+## 0.4.2
+
+### Patch Changes
+
+- Handle OneTrust's short banner-state delay after accepting all consent from either the Contentpass layer or the OneTrust preference center.
+
 ## 0.4.1
 
 ### Patch Changes